Updated quit.sh flow control

+++ scripts/quit.sh
@@ -2,8 +2,12 @@
 
 cd data
 
-rm bookData.txt
-rm tagData.txt
+if [ -f bookData.txt ]; then
+    rm bookData.txt
+fi
+if [ -f tagData.txt ]; then
+    rm tagData.txt
+fi
 
 cat bookOutput.txt | tee bookData.txt
 cat tagOutput.txt | tee tagData.txt
